requirements.txt
django>=3.0,<4.0
djangorestframework>=3.0,<4.0
psycopg2>=2.8